cluster: Move archival_metadata_stm to archival
This commit doesn't change any code. The archival and cluster modules
are still built together by one CMakeLists.txt.
