+++ src/v/archival/archival_metadata_stm.h
